UPSTREAM: audit: add tty field to LOGIN event
The tty field was missing from AUDIT_LOGIN events.

Refactor code to create a new function audit_get_tty(), using it to
replace the call in audit_log_task_info() and to add it to
audit_log_set_loginuid().  Lock and bump the kref to protect it, adding
audit_put_tty() alias to decrement it.

Signed-off-by: Richard Guy Briggs <rgb@redhat.com>
Signed-off-by: Paul Moore <paul@paul-moore.com>
Link: https://git.kernel.org/linus/db0a6fb5d97afe01fd9c47d37c6daa82d4d4001d
Signed-off-by: Nathan Chancellor <natechancellor@gmail.com>
Signed-off-by: เจียน <jianrzy@naver.com>